Enable the default rules
