+++ controllers/commonservice_controller.go
@@ -18,14 +18,8 @@ package controllers
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"reflect"
 
-	utilyaml "github.com/ghodss/yaml"
-	"github.com/go-logr/logr"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog"
@@ -39,8 +33,6 @@ import (
 	util "github.com/IBM/ibm-common-service-operator/controllers/common"
 	"github.com/IBM/ibm-common-service-operator/controllers/constant"
 	"github.com/IBM/ibm-common-service-operator/controllers/deploy"
-	"github.com/IBM/ibm-common-service-operator/controllers/rules"
-	"github.com/IBM/ibm-common-service-operator/controllers/size"
 )
 
 // CommonServiceReconciler reconciles a CommonService object
@@ -49,7 +41,6 @@ type CommonServiceReconciler struct {
 	client.Reader
 	*deploy.Manager
 	*bootstrap.Bootstrap
-	Log    logr.Logger
 	Scheme *runtime.Scheme
 }
 
@@ -220,404 +211,6 @@ func (r *CommonServiceReconciler) ReconcileGeneralCR(instance *apiv3.CommonServi
 
 	klog.Infof("Finished reconciling CommonService: %s/%s", instance.Namespace, instance.Name)
 	return ctrl.Result{}, nil
-}
-
-func (r *CommonServiceReconciler) getNewConfigs(cs *unstructured.Unstructured) ([]interface{}, error) {
-	var newConfigs []interface{}
-	switch cs.Object["spec"].(map[string]interface{})["size"] {
-	case "small":
-		newConfigs, err := applySizeTemplate(cs, size.Small)
-		if err != nil {
-			return newConfigs, err
-		}
-		return newConfigs, nil
-	case "medium":
-		newConfigs, err := applySizeTemplate(cs, size.Medium)
-		if err != nil {
-			return newConfigs, err
-		}
-		return newConfigs, nil
-	case "large":
-		newConfigs, err := applySizeTemplate(cs, size.Large)
-		if err != nil {
-			return newConfigs, err
-		}
-		return newConfigs, nil
-	default:
-		if cs.Object["spec"].(map[string]interface{})["services"] != nil {
-			newConfigs = cs.Object["spec"].(map[string]interface{})["services"].([]interface{})
-		}
-		return newConfigs, nil
-	}
-}
-
-func applySizeTemplate(cs *unstructured.Unstructured, sizeTemplate string) ([]interface{}, error) {
-
-	var src []interface{}
-	if cs.Object["spec"].(map[string]interface{})["services"] != nil {
-		src = cs.Object["spec"].(map[string]interface{})["services"].([]interface{})
-	}
-
-	// Convert sizes string to slice
-	sizes, err := convertStringToSlice(sizeTemplate)
-	if err != nil {
-		klog.Errorf("convert size to interface slice: %v", err)
-		return nil, err
-	}
-
-	for _, configSize := range sizes {
-		config := getItemByName(src, configSize.(map[string]interface{})["name"].(string))
-		if config == nil {
-			continue
-		}
-		if configSize == nil {
-			configSize = config
-			continue
-		}
-		for cr, size := range mergeSizeProfile(configSize.(map[string]interface{})["spec"].(map[string]interface{}), config.(map[string]interface{})["spec"].(map[string]interface{})) {
-			configSize.(map[string]interface{})["spec"].(map[string]interface{})[cr] = size
-		}
-	}
-	return sizes, nil
-}
-
-// mergeCRsIntoOperandConfig merges CRs by specific rules
-func mergeCRsIntoOperandConfig(defaultMap map[string]interface{}, changedMap map[string]interface{}, rules map[string]interface{}) map[string]interface{} {
-	// TODO: Apply different rules
-	for key := range changedMap {
-		// Remove the items not from the rules
-		filterChangedMapWithRules(key, changedMap[key], rules[key], changedMap)
-	}
-	for key := range defaultMap {
-		if reflect.DeepEqual(defaultMap[key], changedMap[key]) {
-			continue
-		}
-		// CR overwrites the existing OperandConfig
-		mergeChangedMap(key, defaultMap[key], changedMap[key], changedMap)
-	}
-	return changedMap
-}
-
-// shrinkSize merges CRs by picking the smaller size
-func shrinkSize(defaultMap map[string]interface{}, changedMap map[string]interface{}) map[string]interface{} {
-	//TODO: Only shrink the parameter with `Largest_value` rule
-	for key := range defaultMap {
-		if reflect.DeepEqual(defaultMap[key], changedMap[key]) {
-			continue
-		}
-		mergeChangedMapWithSmallSize(key, defaultMap[key], changedMap[key], defaultMap)
-	}
-	return changedMap
-}
-
-func mergeCSCRs(csSummary, csCR, ruleslice []interface{}) []interface{} {
-	//TODO: Only merge the parameter with `Largest_value` rule
-	for _, operator := range csCR {
-		summaryCR := getItemByName(csSummary, operator.(map[string]interface{})["name"].(string))
-		rules := getItemByName(ruleslice, operator.(map[string]interface{})["name"].(string))
-		if summaryCR == nil {
-			csSummary = append(csSummary, operator)
-			continue
-		}
-		if summaryCR.(map[string]interface{})["spec"] == nil {
-			csSummary = setSpecByName(csSummary, operator.(map[string]interface{})["name"].(string), operator)
-			continue
-		}
-		for cr, spec := range operator.(map[string]interface{})["spec"].(map[string]interface{}) {
-			if summaryCR.(map[string]interface{})["spec"].(map[string]interface{})[cr] == nil {
-				summaryCR.(map[string]interface{})["spec"].(map[string]interface{})[cr] = spec
-				continue
-			}
-			if rules != nil && rules.(map[string]interface{})["spec"] != nil && rules.(map[string]interface{})["spec"].(map[string]interface{})[cr] != nil {
-				ruleForCR := rules.(map[string]interface{})["spec"].(map[string]interface{})[cr].(map[string]interface{})
-				sizeForCR := summaryCR.(map[string]interface{})["spec"].(map[string]interface{})[cr].(map[string]interface{})
-				summaryCR.(map[string]interface{})["spec"].(map[string]interface{})[cr] = mergeCRsIntoOperandConfig(spec.(map[string]interface{}), sizeForCR, ruleForCR)
-			}
-		}
-	}
-	return csSummary
-}
-
-// mergeCRsIntoOperandConfig merges CRs by specific rules
-func mergeCRsIntoOperandConfigWithDefaultRules(defaultMap map[string]interface{}, changedMap map[string]interface{}) map[string]interface{} {
-	// TODO: Apply default rules
-	for key := range defaultMap {
-		if reflect.DeepEqual(defaultMap[key], changedMap[key]) {
-			continue
-		}
-		mergeChangedMap(key, defaultMap[key], changedMap[key], changedMap)
-	}
-	return changedMap
-}
-
-func filterChangedMapWithRules(key string, changedMap interface{}, rules interface{}, finalMap map[string]interface{}) {
-	switch changedMap.(type) {
-	case map[string]interface{}:
-		//Check that the changed map value doesn't contain this map at all and is nil
-		if rules == nil {
-			delete(finalMap, key)
-		} else {
-			if _, ok := rules.(map[string]interface{}); ok {
-				rulesRef := rules.(map[string]interface{})
-				changedMapRef := changedMap.(map[string]interface{})
-				for newKey := range changedMapRef {
-					filterChangedMapWithRules(newKey, changedMapRef[newKey], rulesRef[newKey], finalMap[key].(map[string]interface{}))
-				}
-			} else {
-				delete(finalMap, key)
-			}
-		}
-	default:
-		if rules == nil && changedMap != nil {
-			delete(finalMap, key)
-		}
-	}
-}
-
-func mergeChangedMap(key string, defaultMap interface{}, changedMap interface{}, finalMap map[string]interface{}) {
-	if !reflect.DeepEqual(defaultMap, changedMap) {
-		switch defaultMap := defaultMap.(type) {
-		case map[string]interface{}:
-			//Check that the changed map value doesn't contain this map at all and is nil
-			if changedMap == nil {
-				finalMap[key] = defaultMap
-			} else if _, ok := changedMap.(map[string]interface{}); ok { //Check that the changed map value is also a map[string]interface
-				defaultMapRef := defaultMap
-				changedMapRef := changedMap.(map[string]interface{})
-				for newKey := range defaultMapRef {
-					mergeChangedMap(newKey, defaultMapRef[newKey], changedMapRef[newKey], finalMap[key].(map[string]interface{}))
-				}
-			}
-		default:
-			//Check if the value was set, otherwise set it
-			if changedMap == nil {
-				finalMap[key] = defaultMap
-			} else {
-				finalMap[key], _ = rules.ResourceComparison(defaultMap, changedMap)
-			}
-		}
-	}
-}
-
-func mergeChangedMapWithSmallSize(key string, defaultMap interface{}, changedMap interface{}, finalMap map[string]interface{}) {
-	if !reflect.DeepEqual(defaultMap, changedMap) {
-		switch changedMap.(type) {
-		case map[string]interface{}:
-			if _, ok := defaultMap.(map[string]interface{}); ok {
-				defaultMapRef := defaultMap.(map[string]interface{})
-				changedMapRef := changedMap.(map[string]interface{})
-				for newKey := range changedMapRef {
-					mergeChangedMapWithSmallSize(newKey, changedMapRef[newKey], defaultMapRef[newKey], finalMap[key].(map[string]interface{}))
-				}
-			}
-		default:
-			//Check if the value was set, otherwise set it
-			if changedMap == nil {
-				finalMap[key] = defaultMap
-			} else {
-				_, finalMap[key] = rules.ResourceComparison(defaultMap, changedMap)
-			}
-		}
-	}
-}
-
-// mergeSizeProfile deep merge two configs
-func mergeSizeProfile(defaultMap map[string]interface{}, changedMap map[string]interface{}) map[string]interface{} {
-	for key := range defaultMap {
-		if reflect.DeepEqual(defaultMap[key], changedMap[key]) {
-			continue
-		}
-		deepMergeTwoMaps(key, defaultMap[key], changedMap[key], changedMap)
-	}
-	return changedMap
-}
-
-func deepMergeTwoMaps(key string, defaultMap interface{}, changedMap interface{}, finalMap map[string]interface{}) {
-	switch defaultMap := defaultMap.(type) {
-	case map[string]interface{}:
-		//Check that the changed map value doesn't contain this map at all and is nil
-		if changedMap == nil {
-			finalMap[key] = defaultMap
-		} else if _, ok := changedMap.(map[string]interface{}); ok { //Check that the changed map value is also a map[string]interface
-			defaultMapRef := defaultMap
-			changedMapRef := changedMap.(map[string]interface{})
-			for newKey := range defaultMapRef {
-				deepMergeTwoMaps(newKey, defaultMapRef[newKey], changedMapRef[newKey], finalMap[key].(map[string]interface{}))
-			}
-		}
-	default:
-		//Check if the value was set, otherwise set it
-		if changedMap == nil {
-			finalMap[key] = defaultMap
-		}
-	}
-}
-
-func (r *CommonServiceReconciler) updateOperandConfig(newConfigs []interface{}) error {
-	opcon := util.NewUnstructured("operator.ibm.com", "OperandConfig", "v1alpha1")
-	opconKey := types.NamespacedName{
-		Name:      "common-service",
-		Namespace: constant.MasterNamespace,
-	}
-	if err := r.Reader.Get(ctx, opconKey, opcon); err != nil {
-		klog.Errorf("failed to get OperandConfig %s: %v", opconKey.String(), err)
-		return err
-	}
-
-	opconServices := opcon.Object["spec"].(map[string]interface{})["services"].([]interface{})
-
-	// Convert rules string to slice
-	ruleSlice, err := convertStringToSlice(rules.ConfigurationRules)
-	if err != nil {
-		return err
-	}
-
-	for _, opService := range opconServices {
-		size := getItemByName(newConfigs, opService.(map[string]interface{})["name"].(string))
-		rules := getItemByName(ruleSlice, opService.(map[string]interface{})["name"].(string))
-		if size == nil {
-			continue
-		}
-		for cr, spec := range opService.(map[string]interface{})["spec"].(map[string]interface{}) {
-			if size.(map[string]interface{})["spec"].(map[string]interface{})[cr] == nil {
-				continue
-			}
-			sizeForCR := size.(map[string]interface{})["spec"].(map[string]interface{})[cr].(map[string]interface{})
-			if rules != nil && rules.(map[string]interface{})["spec"] != nil && rules.(map[string]interface{})["spec"].(map[string]interface{})[cr] != nil {
-				ruleForCR := rules.(map[string]interface{})["spec"].(map[string]interface{})[cr].(map[string]interface{})
-				opService.(map[string]interface{})["spec"].(map[string]interface{})[cr] = mergeCRsIntoOperandConfig(spec.(map[string]interface{}), sizeForCR, ruleForCR)
-			} else {
-				opService.(map[string]interface{})["spec"].(map[string]interface{})[cr] = mergeCRsIntoOperandConfigWithDefaultRules(spec.(map[string]interface{}), sizeForCR)
-			}
-		}
-	}
-
-	// Checking all the common service CRs to get the minimal size
-	opconServices, err = r.getMinimalSizes(opconServices, ruleSlice)
-	if err != nil {
-		return err
-	}
-
-	opcon.Object["spec"].(map[string]interface{})["services"] = opconServices
-
-	if err := r.Update(ctx, opcon); err != nil {
-		klog.Errorf("failed to update OperandConfig %s: %v", opconKey.String(), err)
-		return err
-	}
-
-	return nil
-}
-
-func (r *CommonServiceReconciler) getMinimalSizes(opconServices, ruleSlice []interface{}) ([]interface{}, error) {
-	// Fetch all the CommonService instances
-	csList := util.NewUnstructuredList("operator.ibm.com", "CommonService", "v3")
-	if err := r.Client.List(ctx, csList); err != nil {
-		return []interface{}{}, err
-	}
-	var configSummary []interface{}
-	for _, cs := range csList.Items {
-		if cs.Object["metadata"].(map[string]interface{})["deletionTimestamp"] != nil {
-			continue
-		}
-		csConfigs, err := r.getNewConfigs(&cs)
-		if err != nil {
-			return []interface{}{}, err
-		}
-		configSummary = mergeCSCRs(configSummary, csConfigs, ruleSlice)
-	}
-
-	for _, opService := range opconServices {
-		crSummary := getItemByName(configSummary, opService.(map[string]interface{})["name"].(string))
-		for cr, spec := range opService.(map[string]interface{})["spec"].(map[string]interface{}) {
-			if crSummary == nil || crSummary.(map[string]interface{})["spec"] == nil || crSummary.(map[string]interface{})["spec"].(map[string]interface{})[cr] == nil {
-				continue
-			}
-			serviceForCR := crSummary.(map[string]interface{})["spec"].(map[string]interface{})[cr].(map[string]interface{})
-			opService.(map[string]interface{})["spec"].(map[string]interface{})[cr] = shrinkSize(spec.(map[string]interface{}), serviceForCR)
-		}
-	}
-	return opconServices, nil
-}
-
-func (r *CommonServiceReconciler) handleDelete() error {
-	opcon := util.NewUnstructured("operator.ibm.com", "OperandConfig", "v1alpha1")
-	opconKey := types.NamespacedName{
-		Name:      "common-service",
-		Namespace: constant.MasterNamespace,
-	}
-	if err := r.Reader.Get(ctx, opconKey, opcon); err != nil {
-		klog.Errorf("failed to get OperandConfig %s: %v", opconKey.String(), err)
-		return err
-	}
-
-	opconServices := opcon.Object["spec"].(map[string]interface{})["services"].([]interface{})
-
-	// Convert rules string to slice
-	ruleSlice, err := convertStringToSlice(rules.ConfigurationRules)
-	if err != nil {
-		return err
-	}
-	opconServices, err = r.getMinimalSizes(opconServices, ruleSlice)
-	if err != nil {
-		return err
-	}
-
-	opcon.Object["spec"].(map[string]interface{})["services"] = opconServices
-
-	if err := r.Update(ctx, opcon); err != nil {
-		klog.Errorf("failed to update OperandConfig %s: %v", opconKey.String(), err)
-		return err
-	}
-
-	return nil
-}
-
-func convertStringToSlice(str string) ([]interface{}, error) {
-
-	jsonSpec, err := utilyaml.YAMLToJSON([]byte(str))
-	if err != nil {
-		return nil, fmt.Errorf("failed to convert yaml to json: %v", err)
-	}
-
-	// Create a slice
-	var slice []interface{}
-	// Convert sizes string to slice
-	err = json.Unmarshal(jsonSpec, &slice)
-	if err != nil {
-		return nil, fmt.Errorf("failed to convert string to slice: %v", err)
-	}
-
-	return slice, nil
-}
-
-func getItemByName(slice []interface{}, name string) interface{} {
-	for _, item := range slice {
-		if item.(map[string]interface{})["name"].(string) == name {
-			return item
-		}
-	}
-	return nil
-}
-
-func setSpecByName(slice []interface{}, name string, spec interface{}) []interface{} {
-	for _, item := range slice {
-		if item.(map[string]interface{})["name"].(string) == name {
-			item.(map[string]interface{})["spec"] = spec
-			return slice
-		}
-	}
-	return slice
-}
-
-// Check if the request's NamespacedName is the "master" CR
-func checkNamespace(key string) bool {
-	return key == constant.MasterNamespace+"/common-service"
-}
-
-// updatePhase sets the current Phase status.
-func (r *CommonServiceReconciler) updatePhase(instance *apiv3.CommonService, status string) error {
-	instance.Status.Phase = status
-	return r.Client.Status().Update(ctx, instance)
 }
 
 func (r *CommonServiceReconciler) addFinalizer(instance *apiv3.CommonService) error {
